Move useTabBArHeight hook
